Remove redis and update production config
