+++ docker-compose.yml
@@ -67,13 +67,6 @@ services:
     volumes:
       - ./elasticsearch:/usr/share/elasticsearch/data
     container_name: csx_elastic
-  redis:
-    image: redis:7.0.4
-    ports:
-      - 6379:6379
-    volumes:
-      - ./redis:/data
-    container_name: csx_redis
   mongo:
     image: mongo:4
     ports:
