+++ server/app/routes/search.py
@@ -11,7 +11,6 @@ from pydantic import BaseModel
 
 import app.services.graph.graph as csx_graph
 import app.services.data.elastic as csx_es
-import app.services.data.redis as csx_redis
 import app.services.graph.graph as csx_graph
 import app.services.graph.nodes as csx_nodes
 import app.services.data.autocomplete as csx_auto
@@ -73,190 +72,6 @@ class Data(BaseModel):
     visible_entries: List
     anchor_properties: List
     action_time: str
-
-
-@router.post("/")
-def search(data: Data) -> dict:
-
-    query = data.query
-    user_id = data.user_id
-    search_uuid = data.search_uuid
-    visible_dimensions = data.visible_dimensions
-    schema = data.graph_schema
-    index = data.index
-    anchor = data.anchor
-    links = data.links
-    graph_type = data.graph_type
-    visible_entries = data.visible_entries
-    anchor_properties = data.anchor_properties
-    study_id = data.study_id
-    action_time = data.action_time
-
-    print("\n\n\n\n\n time of search: ", action_time)
-
-    """Run search using given query."""
-    cache_data = csx_redis.load_current_graph(user_id)
-
-    directory_path = os.getcwd()
-    print("My current directory is : " + directory_path)
-
-    # Update study with index
-
-    csx_study.add_index(study_id, user_id, index)
-
-    with open(f"./app/data/config/{index}.json") as config:
-        config = json.load(config)
-
-        dimension_types = config["dimension_types"]
-
-        if len(schema) == 0:
-            schema = config["schemas"][0]["relations"]
-
-        if len(visible_dimensions) == 0:
-            visible_dimensions = config["default_visible_dimensions"]
-
-        default_search_fields = config["default_search_fields"]
-
-        if anchor == "":
-            anchor = config["anchor"]
-
-        if not links:
-            links = config["links"]
-
-    search = Search(using=es, index=index)
-    search = search[0:10000]
-
-    id_list = visible_entries
-    query_generated_dimensions = {}
-
-    if len(id_list):
-        results = convert_filter_res_to_df(
-            search.filter("terms", _id=id_list).execute()
-        )
-    elif not isJson(query) or isNumber(query):
-
-        filtered_fields = default_search_fields
-
-        if not isNumber(query):
-            filtered_fields = [
-                field
-                for field in filtered_fields
-                if dimension_types[field] not in ["integer", "float"]
-            ]
-
-        es_query = Q(
-            "query_string",
-            query=csx_es.convert_to_elastic_safe_query(f"{query}"),
-            type="phrase",
-            fields=filtered_fields,
-        )
-        results = csx_es.query_to_dataframe(es_query, index)
-    else:
-
-        query_generated_dimensions = {
-            entry["feature"]: entry["type"]
-            for entry in get_new_features(json.loads(query))
-        }
-
-        results = csx_es.run_advanced_query(json.loads(query), index, dimension_types)
-
-    if len(results.index) == 0:
-        return {"nodes": []}
-
-    elastic_json = json.loads(results.to_json(orient="records"))
-
-    dimensions = {
-        "links": links,
-        "anchor": {"dimension": anchor, "props": anchor_properties},
-        "visible": visible_dimensions,
-        "query_generated": query_generated_dimensions,
-        "all": [
-            property
-            for property in csx_es.get_index(index)[index]["mappings"]["properties"]
-        ],
-    }
-
-    current_dimensions = visible_dimensions
-
-    if graph_type == "overview":
-        current_dimensions = links + [anchor]
-
-    comparison_res = csx_redis.compare_instances(
-        cache_data,
-        {
-            "index": index,
-            "search_uuid": search_uuid,
-            "query": query,
-            "schema": schema,
-            "dimensions": current_dimensions,
-            "anchor_properties": anchor_properties,
-        },
-        graph_type,
-    )
-
-    comparison_switch = {
-        "from_scratch": lambda: csx_graph.get_graph_from_scratch(
-            graph_type,
-            dimensions,
-            elastic_json,
-            visible_entries,
-            index,
-            cache_data,
-            search_uuid,
-            results,
-            user_id,
-            schema,
-            anchor_properties,
-            comparison_res,
-            study_id,
-            comparison_res["action"],
-            query,
-            action_time,
-            comparison_res["history_action"],
-        ),
-        "from_anchor_properties": lambda: csx_graph.get_graph_with_new_anchor_props(
-            comparison_res,
-            graph_type,
-            dimensions,
-            elastic_json,
-            user_id,
-            study_id,
-            comparison_res["action"],
-            query,
-            index,
-            action_time,
-            comparison_res["history_action"],
-        ),
-        "from_existing_data": lambda: csx_graph.get_graph_from_existing_data(
-            graph_type,
-            dimensions,
-            elastic_json,
-            visible_entries,
-            cache_data,
-            user_id,
-            schema,
-            anchor_properties,
-            index,
-            study_id,
-            comparison_res["action"],
-            query,
-            action_time,
-            comparison_res["history_action"],
-        ),
-        "from_cache": lambda: csx_graph.get_graph_from_cache(
-            comparison_res,
-            graph_type,
-            study_id,
-            comparison_res["action"],
-            query,
-            user_id,
-            index,
-            action_time,
-            comparison_res["history_action"],
-        ),
-    }
-
-    return comparison_switch[comparison_res["action"]]()
 
 
 @router.get("/datasets")
