+++ server/requirements.txt
@@ -12,7 +12,6 @@ wheel
 spacy==3.0.7
 pytextrank
 python-multipart
-redis
 pymongo
 marisa-trie
 palettable
